Homescreen + navbar: Folder Reorganization
