+++ frontend/src/create-quiz/CreateQuizView.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import Header from "../styled-components/header/Header";
+import Header from "../headers/header/Header";
 import CreateQuizQuestionCount from "./CreateQuizQuestionCount";
 import axios from "axios";
 import QuestionCount from "./QuestionCount";
