+++ frontend/src/make-quiz/MakeQuizView.tsx
@@ -2,7 +2,7 @@ import axios from "axios";
 import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import Quiz from "../create-quiz/Quiz";
-import Header from "../styled-components/header/Header";
+import Header from "../headers/header/Header";
 import MakeQuizList from "./MakeQuizList";
 
 const MakeQuizView = () => {
